♻️ Move WorkflowActionsUseCase to RunnerBarCore
- Relocate WorkflowActionsUseCase.swift from Sources/RunnerBar/UseCases/
  to Sources/RunnerBarCore/UseCases/ (mirrors FailureHookRunnerUseCase)
- Update module header comment RunnerBar → RunnerBarCore
- Drop `import RunnerBarCore` (now in the same module)
- Make type and all members public so the app target can still access them
- Delete the old app-target copy

Closes #1594
